recycle hub online fix issue update problem

+++ resources/views/livewire/table/travel-permit.blade.php
@@ -27,11 +27,12 @@
                     <td>{{ $data->address }}</td>
                     <td class="whitespace-no-wrap row-action--icon">
 {{--                        @if(config('app.name', 'Laravel')=='Laravel')--}}
+                        @if(auth()->user()->role==1)
                             <a href="{{ route('travel-permit.show',$data->id) }}" class="btn btn-primary m-1">Detail</a>
                             <a href="{{ route('travel-permit.edit',$data->id) }}" class="btn btn-secondary m-1">Ubah</a>
                             <a href="#" wire:click="deleteItem({{ $data->id }})"
                                class="btn btn-danger m-1 col">Hapus</a>
-{{--                        @endif--}}
+                        @endif
                         <a href="{{ route('travel-permit.download',$data->id) }}" class="btn btn-success m-1">Unduh</a>
 
                     </td>
